TIP-842: Move attribute option model to Akeneo\Pim\Structure
